wip: first pass for a unified Runtime interface, doesn't even compile
