+++ oci/runtimes/kata/kata.go
@@ -1,0 +1,5 @@
+package kata
+
+// TODO
+// This file is pretty much the same as the runc runtime
+// see what you guys can do about it since we now have runtime handlers in Kube
